move: 도메인 영역 분리
