+++ src/main/java/com/example/team1_be/domain/Schedule/Recommend/WeeklySchedule/RecommendedWeeklyScheduleRepository.java
@@ -1,6 +1,5 @@
-package com.example.team1_be.domain.Schedule.Recommend;
+package com.example.team1_be.domain.Schedule.Recommend.WeeklySchedule;
 
-import com.example.team1_be.domain.User.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
