+++ src/main/java/com/example/team1_be/domain/Schedule/Recommend/WorktimeApply/RecommendedWorktimeApply.java
@@ -1,15 +1,13 @@
-package com.example.team1_be.domain.Schedule.Recommend;
+package com.example.team1_be.domain.Schedule.Recommend.WorktimeApply;
 
 import com.example.team1_be.domain.Apply.Apply;
-import com.example.team1_be.domain.Worktime.Worktime;
+import com.example.team1_be.domain.Schedule.Recommend.WeeklySchedule.RecommendedWeeklySchedule;
 import com.example.team1_be.utils.audit.BaseEntity;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
-import javax.validation.constraints.NotNull;
-import java.util.List;
 
 @Getter
 @NoArgsConstructor
